+++ luxai_s2/luxai_s2/unit.py
@@ -72,7 +72,7 @@ class Unit:
         if action.repeat == -1:
             self.action_queue.append(action)
     def move_power_cost(self, rubble_at_target: int):
-        return math.ceil(self.unit_cfg.MOVE_COST + self.unit_cfg.RUBBLE_MOVEMENT_COST * rubble_at_target)
+        return math.floor(self.unit_cfg.MOVE_COST + self.unit_cfg.RUBBLE_MOVEMENT_COST * rubble_at_target)
     def state_dict(self):
         return dict(
             team_id=self.team_id,